backend_old/database.py
from sqlalchemy.ext.asyncio import create_async_engine, AsyncSession, async_sessionmaker
from sqlalchemy.orm import declarative_base
from dotenv import load_dotenv
import os
load_dotenv()

DATABASE_URL = os.getenv(
    "DATABASE_URL",
    "mysql+aiomysql://root:190705@127.0.0.1:3306/medischedule"
)

# Async engine & session
engine = create_async_engine(DATABASE_URL, echo=False)
AsyncSessionLocal = async_sessionmaker(bind=engine, class_=AsyncSession, expire_on_commit=False)

# Base cho ORM
Base = declarative_base()

# Dependency lấy session
async def get_db():
    async with AsyncSessionLocal() as session:
        yield session
