backend_old/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.1
python-dotenv==1.0.0
pydantic>=2.10.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
requests==2.31.0
openai>=1.0.0
annotated-doc==0.0.2
httpx==0.28.1
httpcore==1.0.9
distro==1.9.0
sqlalchemy==2.0.23
aiomysql==0.2.0
mysql-connector-python==8.2.0
PyMySQL==1.1.0
cryptography==41.0.7
greenlet==3.0.3
aiofiles>=23.2.1
google-generativeai>=0.3.0
